+++ packages/atoa_core/lib/src/endpoints/endpoints.dart
@@ -16,5 +16,5 @@ class Endpoints {
   static const getPaymentDetails = 'payments/get-payment-details';
 
   /// Payment Status
-  static String getPaymentStatus(String id) => 'payments/payment-status/$id';
+  static String getPaymentStatus(String id) => 'payments/v2/payment/$id/status';
 }